[CDEC-457] Reunite orphans from Pos.Aeson.Core.Configuration
