+++ core/cardano-sl-core.cabal
@@ -20,7 +20,6 @@ library
   hs-source-dirs:      src
   exposed-modules:
                        Pos.Aeson.Core
-                       Pos.Aeson.Core.Configuration
                        Pos.Aeson.Fee
 
                        Pos.Core
